Move event buffer and event filter interceptor into event-listener component
