+++ components/event/event-listener/src/main/java/uk/gov/justice/services/event/filter/EventFilterInterceptor.java
@@ -1,6 +1,7 @@
-package uk.gov.justice.services.core.eventfilter;
+package uk.gov.justice.services.event.filter;
 
 
+import uk.gov.justice.services.core.eventfilter.EventFilter;
 import uk.gov.justice.services.core.interceptor.Interceptor;
 import uk.gov.justice.services.core.interceptor.InterceptorChain;
 import uk.gov.justice.services.core.interceptor.InterceptorContext;
