+++ components/event/event-listener/src/test/java/uk/gov/justice/services/event/buffer/it/util/repository/StreamBufferOpenEjbAwareJdbcRepository.java
@@ -1,8 +1,8 @@
-package uk.gov.justice.services.core.it.util.repository;
+package uk.gov.justice.services.event.buffer.it.util.repository;
 
 import uk.gov.justice.services.event.buffer.core.repository.streambuffer.StreamBufferJdbcRepository;
 
-public class StreamBufferOpenEjbAwareJdbcRepository extends StreamBufferJdbcRepository{
+public class StreamBufferOpenEjbAwareJdbcRepository extends StreamBufferJdbcRepository {
     @Override
     protected String jndiName() {
         return "java:openejb/Resource/viewStore";
