+++ core/src/test/java/uk/gov/justice/services/core/it/MetricsIT.java
@@ -28,22 +28,18 @@ import uk.gov.justice.services.core.dispatcher.RequesterProducer;
 import uk.gov.justice.services.core.dispatcher.ServiceComponentObserver;
 import uk.gov.justice.services.core.dispatcher.SystemUserUtil;
 import uk.gov.justice.services.core.enveloper.Enveloper;
-import uk.gov.justice.services.core.eventfilter.AllowAllEventFilter;
 import uk.gov.justice.services.core.extension.AnnotationScanner;
 import uk.gov.justice.services.core.extension.BeanInstantiater;
 import uk.gov.justice.services.core.interceptor.InterceptorCache;
 import uk.gov.justice.services.core.interceptor.InterceptorChainProcessor;
 import uk.gov.justice.services.core.interceptor.InterceptorChainProcessorProducer;
 import uk.gov.justice.services.core.interceptor.InterceptorObserver;
-import uk.gov.justice.services.core.it.util.repository.StreamBufferOpenEjbAwareJdbcRepository;
-import uk.gov.justice.services.core.it.util.repository.StreamStatusOpenEjbAwareJdbcRepository;
 import uk.gov.justice.services.core.jms.DefaultJmsDestinations;
 import uk.gov.justice.services.core.jms.JmsSenderFactory;
 import uk.gov.justice.services.core.metrics.MetricRegistryProducer;
 import uk.gov.justice.services.core.metrics.MetricsInterceptor;
 import uk.gov.justice.services.core.sender.ComponentDestination;
 import uk.gov.justice.services.core.sender.SenderProducer;
-import uk.gov.justice.services.event.buffer.core.service.ConsecutiveEventBufferService;
 import uk.gov.justice.services.messaging.JsonEnvelope;
 import uk.gov.justice.services.messaging.JsonObjectEnvelopeConverter;
 import uk.gov.justice.services.messaging.jms.DefaultJmsEnvelopeSender;
@@ -87,7 +83,6 @@ public class MetricsIT {
             InterceptorChainProcessor.class,
             InterceptorCache.class,
             InterceptorObserver.class,
-            EventHandlerIT.TestInterceptor.class,
 
             SenderProducer.class,
             JmsSenderFactory.class,
@@ -108,11 +103,7 @@ public class MetricsIT {
             DispatcherCache.class,
             DispatcherFactory.class,
             PolicyEvaluator.class,
-            AllowAllEventFilter.class,
 
-            StreamBufferOpenEjbAwareJdbcRepository.class,
-            StreamStatusOpenEjbAwareJdbcRepository.class,
-            ConsecutiveEventBufferService.class,
             LoggerProducer.class,
             EmptySystemUserProvider.class,
             BeanInstantiater.class,
