+++ generators/rest-client-generator/src/test/java/uk/gov/justice/api/RemoteExampleEventProcessorIT.java
@@ -34,7 +34,6 @@ import uk.gov.justice.services.core.dispatcher.RequesterProducer;
 import uk.gov.justice.services.core.dispatcher.ServiceComponentObserver;
 import uk.gov.justice.services.core.dispatcher.SystemUserUtil;
 import uk.gov.justice.services.core.enveloper.Enveloper;
-import uk.gov.justice.services.core.eventbuffer.PassThroughEventBufferService;
 import uk.gov.justice.services.core.extension.BeanInstantiater;
 import uk.gov.justice.services.core.interceptor.InterceptorCache;
 import uk.gov.justice.services.core.interceptor.InterceptorChainProcessor;
@@ -132,7 +131,6 @@ public class RemoteExampleEventProcessorIT {
             DispatcherFactory.class,
             PolicyEvaluator.class,
 
-            PassThroughEventBufferService.class,
             LoggerProducer.class,
             EmptySystemUserProvider.class,
             SystemUserUtil.class,
