+++ test-utils/test-utils-persistence/pom.xml
@@ -57,11 +57,6 @@
             <artifactId>mockito-core</artifactId>
             <scope>test</scope>
         </dependency>
-        <dependency>
-            <groupId>uk.gov.justice.services</groupId>
-            <artifactId>test-utils-common</artifactId>
-            <version>${project.version}</version>
-        </dependency>
     </dependencies>
